created status light library
